Extract UBlox into custom UserControl
